+++ examples/transaction/transfer_transaction_fungible.py
@@ -56,7 +56,7 @@ def create_account(client, operator_key):
     try:
         tx = (
             AccountCreateTransaction()
-            .set_key(recipient_key.public_key())
+            .set_key_without_alias(recipient_key.public_key())
             .set_initial_balance(Hbar.from_tinybars(100_000_000))
         )
         receipt = tx.freeze_with(client).sign(operator_key).execute(client)