+++ tests/integration/token_revoke_kyc_transaction_e2e_test.py
@@ -20,7 +20,7 @@ def test_token_revoke_kyc_transaction_can_execute():
         # Create a new account
         receipt = (
             AccountCreateTransaction()
-            .set_key(new_account_public_key)
+            .set_key_without_alias(new_account_public_key)
             .set_initial_balance(Hbar(2))
             .execute(env.client)
         )
@@ -72,7 +72,7 @@ def test_token_revoke_kyc_transaction_fails_with_no_kyc_key():
         # Create a new account
         receipt = (
             AccountCreateTransaction()
-            .set_key(new_account_public_key)
+            .set_key_without_alias(new_account_public_key)
             .set_initial_balance(Hbar(2))
             .execute(env.client)
         )
@@ -124,7 +124,7 @@ def test_token_revoke_kyc_transaction_fails_when_account_not_associated():
         # Create a new account
         receipt = (
             AccountCreateTransaction()
-            .set_key(new_account_public_key)
+            .set_key_without_alias(new_account_public_key)
             .set_initial_balance(Hbar(2))
             .execute(env.client)
         )